ContentLibSubsystem.cpp: Simplify getting recipe producers
`FGRecipe.h` has a static function that returns a `TArray<UObject>` of
a recipe's producers. Since many functions only iterate over the array,
using this static function instead of the member function (which takes
the `TArray` as an in/out parameter) saves a few lines of code.

This change also avoids getting the CDO, then its class, then the CDO,
which is what happened before: the member function takes the CDO, but
then calls the static function with `CDO->GetClass()` as parameter.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>